+++ app/src/main/java/woowacourse/shopping/domain/repository/ShoppingCartRepository.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.domain.repository
 
-import woowacourse.shopping.domain.model.ShoppingCartItem
+import woowacourse.shopping.domain.model.shoppingcart.ShoppingCartItem
 
 interface ShoppingCartRepository {
     fun saveItem(
