+++ app/src/main/java/woowacourse/shopping/presentation/shoppingcart/ShoppingCartClickListener.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.presentation.shoppingcart
 
-import woowacourse.shopping.domain.model.ShoppingCartItem
+import woowacourse.shopping.domain.model.shoppingcart.ShoppingCartItem
 
 fun interface ShoppingCartClickListener {
     fun onDeleteGoods(item: ShoppingCartItem)
