+++ app/src/main/java/woowacourse/shopping/presentation/shoppingcart/ShoppingCartViewHolder.kt
@@ -5,7 +5,7 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.shopping.R
 import woowacourse.shopping.databinding.ItemShoppingCartBinding
-import woowacourse.shopping.domain.model.ShoppingCartItem
+import woowacourse.shopping.domain.model.shoppingcart.ShoppingCartItem
 import woowacourse.shopping.presentation.util.QuantityClickListener
 
 class ShoppingCartViewHolder(
