+++ app/src/main/java/woowacourse/shopping/presentation/util/QuantityClickListener.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.presentation.util
 
-import woowacourse.shopping.domain.model.ShoppingCartItem
+import woowacourse.shopping.domain.model.shoppingcart.ShoppingCartItem
 
 interface QuantityClickListener {
     fun increase(item: ShoppingCartItem)
