+++ app/src/main/res/layout/item_goods.xml
@@ -9,7 +9,7 @@
 
         <variable
             name="item"
-            type="woowacourse.shopping.domain.model.ShoppingCartItem" />
+            type="woowacourse.shopping.domain.model.shoppingcart.ShoppingCartItem" />
 
         <variable
             name="quantityClickListener"
