+++ app/src/main/res/layout/view_quantity_controller.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="item"
-            type="woowacourse.shopping.domain.model.ShoppingCartItem" />
+            type="woowacourse.shopping.domain.model.shoppingcart.ShoppingCartItem" />
 
         <variable
             name="quantityClickListener"
